Speed up timeline requests (#5)
Initially, we attempted to speed up timeline queries by streaming them instead of making HTTP requests. Half way though, I discovered that the reason the queries were taking so long was because the whole database was being returned and we were incldued a bunch of unnecessary data (event.data) in responses that took a while to json serialize.

So now we have the foundation of timeline streaming, and more performant queries.